requirements.txt
absl-py==0.9.0
aiohttp==3.6.3
aniso8601==8.0.0
appdirs==1.4.4
appnope==0.1.0
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==1.5.0
cachetools==4.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
components==1.2.8
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
entrypoints==0.3
enum34==1.1.10
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
flask-cors==3.0.9
future==0.18.2
gast==0.3.3
gcloud==0.18.3
geoip2==4.1.0
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
heroku==0.1.4
html5lib==0.9999999
httplib2==0.18.1
idna==2.9
idna-ssl==1.1.0
imageio==2.9.0
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.17.0
JPype1==1.1.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
konlpy==0.5.2
lxml==4.6.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
maxminddb==2.0.3
mistune==0.8.4
multidict==4.7.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
npm==0.1.1
numpy==1.18.2
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.2.0
optional-django==0.1.0
packaging==20.4
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.0
path==15.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.23.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
style==1.1.0
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
timezones==2.0.7
tornado==6.0.4
traitlets==4.3.3
tweepy==3.9.0
typing-extensions==3.7.4.3
update==0.0.1
urllib3==1.25.8
uWSGI==2.0.19.1
virtualenv==20.0.21
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.5.1
zipp==3.1.0
